Delete Appointment and setup with error controls to restrict user from deleting if customer is tied to appointment.

+++ Software2/Repository/CustomerRepository.cs
@@ -64,6 +64,10 @@ namespace Software2.Forms
         internal void deleteCustomer(int customerID)
         {
             var customerToDelete = getCustomerByID(customerID);
+            if(customerToDelete.appointments.Count > 0)
+            {
+                throw new Exception("Cannot delete customer that is attached to an appointment");
+            }
             db.customers.Remove(customerToDelete);
             db.SaveChanges();
         }
